requirements.txt
Flask==2.3.2
transformers==4.57.1
torch==2.9.0
Pillow==12.0.0
python-dotenv==1.1.1
huggingface-hub==0.36.0
